Add code to initialize RPC and DP communication
The plugin will communicate with the dataplane over a unix socket
with a proprietary binary RPC format. hh_dp_comm.c|h includes code
to open a unix socket and bind it to a fixed location in the file-
system. The code will also connect the socket to the dataplane at
a fixed location. If the connection fails, it will be re-attempted
after a fixed amount of time. Neither the paths to bind/connect
the unix socket nor the reconnection interval are configurable yet.