Define dp_msg object and add message cache
dp_msg objects are a wrapper to RpcMsg objects defined by the
wire protocol that include a pointer to the ctx object queued by
zebra. For every request, a dp_msg object has to be created and let
it refer to the original ctx queued by zebra.

This patch also defines a message cache entity which contains 3
lists or queues. A "pool" list contains unused dp_msg objects that
are free for use. This is just a pool of blank dp_msg objects to
avoid allocating dp_msg's on every request. When no longer used,
dp_msg objects are recycled to the pool. Another list "unsent"
contains dp_msg's that have been populated but not yet sent; e.g.
because the dataplane did not confirm a Connect request yet or
because the socket is momentarily not accepting writes. Lastly,
an "in-flight" list contains dp_msg's that have been sent over the
unix socket and for which a response is expected. Upon reception
of responses, dp_msg's from the in-flight list may be popped,
and the zebra ctx they point to updated and queued back to zebra.
At that point, dp_msg's may be recycled back into the "pool" list.