Fix: Correct issuing Connect requests
Once the plugin is registered and started, zebra will queue
requests to it, whether the plugin has contacted or not the data-
plane. In order not to lose requests, these get cached and not sent
until the plugin succesfully connects the unix socket and receives
a positive response to a Connect request. In the prior code,
connects were treated as regular requests. Hence if prior requests
had been made, connect requests were cached afterwards and never
sent. To fix this:
 - don't cache Connect requests in the unsent queue, but rather
   send them and cache them in the in-flight queue (which should be
   empty)
-  let connect requests have always zero sequence number / Id. This
   is mostly cosmetic so that in logs we don't see that Connect
   request #34 succeded and then about 33 other requests (that
   were cached) be sent.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>