Add structs and functions to store RPC stats.
We're mostly concerned with Requests and responses at this point.

Since the RPC is ~ CRUD, it makes sense to keep statistics for the
number of operations requested (e.g. Add/Del/Update...) on a per
object type basis (e.g. routes, rmacs, etc..). For each tuple
(operation, object-type), we may store: the number of requests
sent, number of responses received and whether they failed or not,
in some sort of stats cell or record. One way to implement this
is with a hash table of stats cells keyed by the tuple above.
A simpler, more efficient one is to keep a 2-dimensional matrix
of stats cells, as this requires a single allocation (in stack).
Also, we use atomic types for the counters since those may be read
from a distinct thread. This should only matter under heavy RPC
activity.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>