utils: add utils to parse args to the plugin.
This plugin will require some configuration parameters.
The FRR module subsystem has a mechanism to propagate parameters to
modules. However, these come packed as a single string that has to
be tokenized. Also, frr_getopt() in libfrr cannot be used as it has
hardwired options and handlers which may differ from ours.

This patch adds generic tooling to parse a set of getopt-like
parameters concatenated as a single string. These utils accept a
(specific) callback to determine the actions to perform when the
specified options are encountered and have no hardwired set of
options. The utils expect (and work) with both short and long
options. E.g.
  short:
     -e FOO
  long:
     --example FOO or --example=FOO

In the last example using '=' no space should be present between
the option and the value.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>