Add vtysh extension module
Create module to be dynamically loaded by vtysh to be able to
issue commands to the Hedgehog gateway plugin. Using a dynamically
loaded module allows us to build the FRR independently and avoid
a rebuild each time a command is added / changed.

So, this repo now produces 2 objects: the plugin itself, and a
tiny shared library to be dynamically loaded by vtysh. The library
name will not be hardcoded in FRR/vtysh but passed as argument
(e.g. vtysh -X module-name) for flexibility. In order not to have
to specify the module on vtysh invocation, we may define a bash
alias in the FRR container.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>