Handle stale requests on dataplane restart
If dataplane restarts, we may never receive responses to requests
that we issued. When detecting a restart, purge the in-flight
queue (which contains the requests to the prior dataplane instance)
and answer them so that FRR believes they succeeded. This will not
create a mismatch in state between dataplane and frr since the
former will request a refresh, whose outcome (responses) will
set the frr state to the right one.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>