Fix purging logic of prior commit.
While purging, which occurs when we detect that dataplane restarted
by processing a new Connect, do not send pending messages. Else,
they will be moved to the in-flight cache and be purged too: we'll
answer them on behalf of dataplane and, when the real answers
come, we'll not know what to do with them.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>